Move library resolution code into config
